* Allow negative values for ScrollWheelSpeed
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2729 37977431-3df6-0310-b722-df95706aa16b

+++ rts/System/MouseInput.h
@@ -19,7 +19,7 @@ public:
 	virtual void HandleSDLMouseEvent (SDL_Event& event) = 0;
 
 protected:
-	Uint8 scrollWheelSpeed;
+	int scrollWheelSpeed;
 };
